requirements.txt
Django==1.6.5
argparse==1.2.1
dj-database-url==0.2.2
dj-static==0.0.5
django-toolbelt==0.0.1
gunicorn==18.0
psycopg2==2.5.1
static==0.4
wsgiref==0.1.2

BeautifulSoup==3.2.1
Cython==0.20.1
Pillow==2.3.0
django-debug-toolbar==1.2
django-socketio==0.3.9
gevent-socketio==0.2.1
gevent-websocket==0.9.3
django-easy-pdf
git+https://github.com/chrisglass/xhtml2pdf.git
coverage==3.5
html5lib==0.90
httplib2==0.7.6
nose==1.0.0
pyPdf2==1.19
reportlab==2.5
